Traps closing the window to a dialog box
Just for confirmation

+++ src/playerboundary/Application.java
@@ -49,26 +49,36 @@ public class Application extends JFrame {
 	/**
 	 * Default constructor for LetterCraze Player Application.
 	 */
-	
+
 	public Application(Model model) {
 		//pack(); not really sure what pack does
+		
+		// this handles all events pertaining to closing the Application
 		this.addWindowListener(new java.awt.event.WindowAdapter() {
-		    @Override
-		    public void windowClosing(java.awt.event.WindowEvent windowEvent) {
-		    	try {
-		            FileOutputStream fileOut =
-		            new FileOutputStream("playersave.ser");
-		            ObjectOutputStream out = new ObjectOutputStream(fileOut);
-		            out.writeObject(model);
-		            out.close();
-		            fileOut.close();
-		            System.out.printf("Serialized data is saved in playersave.ser");
-		         }catch(IOException i) {
-		            i.printStackTrace();
-		         }
-		    }
+			@Override
+			public void windowClosing(java.awt.event.WindowEvent windowEvent) {
+				int confirmed = JOptionPane.showConfirmDialog(null, 
+						"Are you sure you want to exit LetterCraze?\n"
+						+ "All data will be saved when you exit.", "",
+						JOptionPane.YES_NO_OPTION);
+
+				if (confirmed == JOptionPane.YES_OPTION) {
+					try {
+						FileOutputStream fileOut =
+								new FileOutputStream("playersave.ser");
+						ObjectOutputStream out = new ObjectOutputStream(fileOut);
+						out.writeObject(model);
+						out.close();
+						fileOut.close();
+						System.out.printf("Serialized data is saved in playersave.ser");
+					}catch(IOException i) {
+						i.printStackTrace();
+					}
+					dispose();
+				}
+			}
 		});
-		setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+		setDefaultCloseOperation(JFrame.DO_NOTHING_ON_CLOSE);
 		setBounds(100, 100, 800, 600);
 		setTitle("LetterCraze Player");
 
@@ -130,14 +140,14 @@ public class Application extends JFrame {
 		splashScreenTimer = new Timer(1000, new SplashScreenTimerController(this, this.model));
 		splashScreenTimer.start();
 		this.setContentPane(splashScreen);
-		
+
 
 	}
 
 	/**
 	 * Getter for how many seconds to display Splash Screen.
 	 */
-	
+
 	public int getDisplayTime() {
 		return displayTime;
 	}
@@ -145,20 +155,20 @@ public class Application extends JFrame {
 	/**
 	 * Getter for Timer for determining how long Splash Screen should be displayed.
 	 */
-	
+
 	public Timer getSplashScreenTimer() {
 		return splashScreenTimer;
 	}
-	
+
 	/**
 	 * Getter for decrementing time of Splash Screen.
 	 */
-	
+
 
 	public void decrementDisplayTime() {
 		displayTime--;
 	}
-	
+
 	/**
 	 * Getter for Main Menu GUI after Splash Screen is implemented.
 	 */
@@ -166,7 +176,7 @@ public class Application extends JFrame {
 	public InitializeGameApplication getMainMenu() {
 		return mainMenu;
 	}
-	
+
 	/**
 	 * Getter for Map GUI.
 	 */
@@ -178,7 +188,7 @@ public class Application extends JFrame {
 	/**
 	 * Getter for Custom Map GUI.
 	 */
-	
+
 	public ViewCustomLevelsApplication getViewCustomLevelsApplication() {
 		return customMenu;
 	}
@@ -186,11 +196,11 @@ public class Application extends JFrame {
 	/**
 	 * Getter for Level GUI.
 	 */
-	
+
 	public ArrayList<LevelApplication> getLevelApplications() {
 		return levelApplications;
 	}
-	
+
 	/**
 	 * Getter for Custom Level GUI.
 	 */
@@ -198,11 +208,11 @@ public class Application extends JFrame {
 	public ArrayList<LevelApplication> getCustomLevelApplications() {
 		return customLevelApplications;
 	}
-	
+
 	/**
 	 * Getter for Model Entities which contains adventure mode level information.
 	 */
-	
+
 	public Model getModel() {
 		return model;
 	}
